DDD_Project
The purpose of this project is to develop a system that can accurately identify and alert drivers who are becoming drowsy or fatigued while operating a vehicle. By leveraging machine learning algorithms and real-time data analysis, the system can provide effective and reliable solution to address the critical issue of driver drowsiness, potentially saving lives and reducing the risks of accidents.
Technology Used: - Deep Learning, Python.